pyproject.toml
[project]
name = "docpeel"
version = "0.1.0"
description = ""
authors = [
    {name = "apmariglis",email = "apmariglis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic (>=0.84.0,<0.85.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "numpy (>=2.0)",
    "google-genai (>=1.65.0,<2.0.0)",
    "mistralai (>=1.12.4,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "docpeel", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pdbpp (>=0.12.1,<0.13.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
